requirements.txt
Flask==3.0.2
flask-cors==4.0.0
Werkzeug==3.0.1 

FastAPI==0.111.0
uvicorn==0.24.0.post1
python-multipart==0.0.9 
python-dotenv==1.0.1 
requests==2.31.0
httpx==0.27.0

pydantic==2.11.7
pydantic-settings==2.3.4
